Various improvments for C++ future standard compliance
GIT_ORIGIN_FUTURE_CPP_REV_ID: de7d969c2ba87b9f416c58f35c0c932e6cd18b68

+++ score/language/futurecpp/src/charconv.cpp
@@ -56,7 +56,7 @@ to_chars_result to_chars_base16_impl(char* first, char* last, const T value, con
     }
 
     const string_view digits{"0123456789abcdef"};
-    span<char> hex_string{first, buffer_len};
+    span<char> hex_string{first, static_cast<std::size_t>(buffer_len)};
 
     std::ptrdiff_t index_hex_string{0};
     if (signed_offset > 0)
